casting indexes to int in bbinom, bnbinom, gpois, neghyper

+++ src/beta-binomial-distribution.cpp
@@ -191,9 +191,9 @@ NumericVector cpp_pbbinom(
       Rcpp::warning("NAs introduced by coercion to integer range");
     } else {
       
-      std::vector<double>& tmp = memo[std::make_tuple(i % size.length(),
-                                                      i % alpha.length(),
-                                                      i % beta.length())];
+      std::vector<double>& tmp = memo[std::make_tuple(static_cast<int>(i % size.length()),
+                                                      static_cast<int>(i % alpha.length()),
+                                                      static_cast<int>(i % beta.length()))];
       if (!tmp.size()) {
         tmp = cdf_bbinom_table(mx, GETV(size, i), GETV(alpha, i), GETV(beta, i));
       }
